completion of Add and Read operations

+++ src/components/Pages/Register.js
@@ -1,5 +1,6 @@
 import { React, useState } from "react";
 import { Form, Button, Card, Container, Alert } from "react-bootstrap";
+// import { useHistory } from "react-router";
 import classes from "./Register.module.css";
 
 const Register = () => {
@@ -8,6 +9,8 @@ const Register = () => {
   const [inputConfirmPassword, setInputConfirmPassword] = useState();
   const [isPasswordMatch, setIsPasswordMatch] = useState(true);
   const [isReistered, setIsRegistered] = useState(false);
+  const [isMailIdExist, setIsMailIdExist] = useState(false);
+  //   const history = useHistory();
   const userDetailsArray = [];
   const formSubmitHandler = (event) => {
     event.preventDefault();
@@ -19,10 +22,20 @@ const Register = () => {
         password: inputPassword,
       };
       //   userDetailsArray.push(inputEmail);
+      const isUserExist = JSON.parse(localStorage.getItem("userCredentials"));
+      console.log(isUserExist);
+      if (isUserExist && isUserExist[0].email === inputEmail) {
+        setIsMailIdExist(true);
+        setIsRegistered(false);
+        return;
+      }
+      setIsMailIdExist(false);
+      console.log(isMailIdExist);
       userDetailsArray.push(user);
       localStorage.setItem("userCredentials", JSON.stringify(userDetailsArray));
       console.log(JSON.parse(localStorage.getItem("userCredentials")));
       setIsRegistered(true);
+      //   history.push("/login");
       return;
     }
     console.log("Password does not match");
@@ -43,13 +56,18 @@ const Register = () => {
     console.log(event.target.value);
     setInputConfirmPassword(event.target.value);
   };
-  console.log(isPasswordMatch);
+  //   console.log(isPasswordMatch);
   return (
     <Container>
       <Card className={classes.card}>
         <Card.Body>
           <Card.Title>Sign Up</Card.Title>
           <br />
+          {isMailIdExist && (
+            <Alert className={classes.alertMessage} variant={"danger"}>
+              Mail id already exist
+            </Alert>
+          )}
           {isReistered && (
             <Alert className={classes.alertMessage} variant={"success"}>
               Registered successfully
